Move LoadOptions to data/index
